+++ R/funcGam.R
@@ -38,25 +38,20 @@
 #'   metricRes$patient_stage, "|", metricRes$patient_id,
 #'   "|", metricRes$image_number
 #' )
-#' # prepare data for FDA
-#' dat <- prepData(metricRes, "r", "rs")
+#' dat <- prepData(metricRes, "r", "rs", sample_id = "patient_id",
+#'     image_id = "image_number", condition = "patient_stage")
 #'
-#' # drop rows with NA
+#'#' # drop rows with NA
 #' dat <- dat |> drop_na()
 #'
-#' # create meta info of the IDs
-#' splitData <- strsplit(dat$ID, "|", fixed = TRUE)
-#' dat$condition <- factor(sapply(splitData, function(x) x[1]))
-#' dat$patient_id <- factor(sapply(splitData, function(x) x[2]))
-#' dat$image_id <- factor(sapply(splitData, function(x) x[3]))
 #' # create a designmatrix
-#' condition <- dat$condition
+#' condition <- dat$patient_stage
 #' # relevel the condition - can set explicit contrasts here
 #' condition <- relevel(condition, "Non-diabetic")
 #' designmat <- model.matrix(~condition)
 #' # colnames don't work with the '-' sign
 #' colnames(designmat) <- c(
-#'     "Intercept", "conditionLong_duration",
+#'     "(Intercept)", "conditionLong_duration",
 #'     "conditionOnset"
 #' )
 #' # fit the model
@@ -90,7 +85,7 @@ functionalGam <- function(data, x, designmat, weights, formula,
     # with the exception of the intercept as this is inferred - there are
     # functional and constant intercepts
     # deselect "Intercept"
-    colNames <- colnames(designmat)[!colnames(designmat) == "Intercept"]
+    colNames <- colnames(designmat)[!colnames(designmat) == "(Intercept)"]
     # stop if the rest terms of design mat are not in the formula
     stopifnot(colNames %in% attr(terms(formula), "term.labels"))
     mdl <- refund::pffr(formula,