+++ R/prepData.R
@@ -4,6 +4,9 @@
 #' the columns ID (unique identifier of each row)
 #' @param x the name of the x-axis of the spatial metric
 #' @param y the name of the y-axis of the spatial metric
+#' @param sample_id the spe `colData` variable to mark the sample
+#' @param image_id the spe `colData` variable to mark the image
+#' @param condition the spe `colData` variable to mark the condition
 #'
 #' @return returns a list with three entries, the unique ID, the functional
 #' response Y and the weights
@@ -26,10 +29,12 @@
 #'     metricRes$patient_stage, "|", metricRes$patient_id,
 #'     "|", metricRes$image_number
 #' )
-#' dat <- prepData(metricRes, "r", "rs")
+#' dat <- prepData(metricRes, "r", "rs", sample_id = "patient_id",
+#' image_id = "image_number", condition = "patient_stage")
 #' @import tidyr
 #' @importFrom methods is
-prepData <- function(metricRes, x, y) {
+prepData <- function(metricRes, x, y, sample_id = NULL, image_id = NULL,
+                     condition = NULL){
     # type checking
     stopifnot(is(metricRes, "data.frame"))
     stopifnot(is(x, "character"))
@@ -50,12 +55,15 @@ prepData <- function(metricRes, x, y) {
       unique()
     # add the weights to the data.frame
     dat <- dat %>% dplyr::left_join(weights, by = "ID")
-    # extract the coordinates
-    coords <- metricRes %>%
-        dplyr::select("ID", "centroidx", "centroidy") %>%
-        unique()
+    # extract the coordinates and meta data and convert to factor
+    meta <- metricRes %>%
+        dplyr::select("ID", "centroidx", "centroidy", sample_id,
+                      image_id, condition) %>%
+        unique() %>%
+        mutate(across(c(sample_id,
+                        image_id, condition), as.factor))
     # add the coordinates to the data.frame
-    dat <- dat %>% dplyr::left_join(coords, by = "ID")
+    dat <- dat %>% dplyr::left_join(meta, by = "ID")
 
     return(dat)
 }